Title: 피보나치 속 일의 자리, Time: 177ms, Memory: 11MB, Status: Wrong Answer - CodeTree

+++ 240315/피보나치 속 일의 자리/number-of-places-in-the-fibonacci.java
@@ -1,12 +1,12 @@
 import java.util.*;
 public class Main {
     public static final int MAX_LENGTH = 1000;
-    public static int[] arr = new int[MAX_LENGTH];
+    public static long[] arr = new long[MAX_LENGTH];
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
         int n = sc.nextInt();
-        arr[0] = sc.nextInt();
-        arr[1] = sc.nextInt();
+        arr[0] = sc.nextLong();
+        arr[1] = sc.nextLong();
         for(int i = 2 ; i < n ; i++){
             arr[i] = arr[i - 1] + arr[i - 2];
         }
